requirements.txt
aiodns==3.2.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
idna==3.7
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
load-dotenv==0.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
orjson==3.10.6
packaging==24.1
paramiko==3.4.0
pfzy==0.3.4
pillow==10.4.0
prettytable==3.10.2
prompt_toolkit==3.0.47
py-cpuinfo==9.0.0
pycares==4.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
runpod==1.6.2
s3transfer==0.10.2
safetensors==0.4.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.0
torch==2.3.1
tqdm==4.66.4
tqdm-loggable==0.2
transformers==4.42.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
